refactor auto-instrumentation moving a few things around, in hopes of making it work with php7
