+++ otel/src/auto/plugin.rs
@@ -1,11 +1,4 @@
 use std::sync::Arc;
-use opentelemetry::{
-    KeyValue,
-    trace::{
-        SpanKind,
-        SpanRef,
-    },
-};
 use phper::{
     values::{ExecuteData, ZVal},
     objects::ZObj,
@@ -26,12 +19,12 @@ pub trait Handler: Send + Sync {
 }
 
 pub struct HandlerCallbacks {
-    pub pre_observe: Option<unsafe extern "C" fn(*mut ExecuteData, &mut SpanDetails)>,
-    pub post_observe: Option<unsafe extern "C" fn(*mut ExecuteData, &SpanRef, &mut ZVal, Option<&mut ZObj>)>,
+    pub pre_observe: Option<unsafe extern "C" fn(*mut ExecuteData)>,
+    pub post_observe: Option<unsafe extern "C" fn(*mut ExecuteData, &mut ZVal, Option<&mut ZObj>)>,
 }
 
-pub type ObserverPreHook = Box<dyn Fn(&mut ExecuteData, &mut SpanDetails) + Send + Sync>;
-pub type ObserverPostHook = Box<dyn Fn(&mut ExecuteData, &SpanRef, &mut ZVal, Option<&mut ZObj>) + Send + Sync>;
+pub type ObserverPreHook = Box<dyn Fn(&mut ExecuteData) + Send + Sync>;
+pub type ObserverPostHook = Box<dyn Fn(&mut ExecuteData, &mut ZVal, Option<&mut ZObj>) + Send + Sync>;
 
 pub struct FunctionObserver {
     pre_hooks: Vec<ObserverPreHook>,
@@ -66,36 +59,5 @@ impl FunctionObserver {
     /// Checks if this function has any hooks
     pub fn has_hooks(&self) -> bool {
         !self.pre_hooks.is_empty() || !self.post_hooks.is_empty()
-    }
-}
-
-#[derive(Debug)]
-pub struct SpanDetails {
-    name: String,
-    attributes: Vec<KeyValue>,
-    kind: SpanKind,
-}
-
-impl SpanDetails {
-    pub fn new(name: String, attributes: Vec<KeyValue>) -> Self {
-        Self {name: name.clone(), attributes, kind: SpanKind::Internal }
-    }
-    pub fn add_attribute(&mut self, kv: KeyValue) {
-        self.attributes.push(kv);
-    }
-    pub fn update_name(&mut self, name: &str) {
-        self.name = name.to_string();
-    }
-    pub fn set_kind(&mut self, kind: SpanKind) {
-        self.kind = kind;
-    }
-    pub fn name(&self) -> String {
-        self.name.clone()
-    }
-    pub fn attributes(&self) -> Vec<KeyValue> {
-        self.attributes.clone()
-    }
-    pub fn kind(&self) -> SpanKind {
-        self.kind.clone()
     }
 }
