+++ otel/tests/auto/auto-and-manual.phpt
@@ -39,7 +39,7 @@ Span #1
 	Instrumentation Scope
 		Name         : "php-auto-instrumentation"
 %A
-	Name        : i-was-renamed
+	Name        : demo-function
 	TraceId     : %s
 	SpanId      : %s
 	TraceFlags  : TraceFlags(1)
@@ -52,5 +52,5 @@ Span #1
 		 ->  code.function.name: String(Owned("demoFunction"))
 		 ->  code.file.path: String(Owned("%s"))
 		 ->  code.line.number: I64(%d)
-		 ->  my-attribute: String(Static("my-value"))
+		 ->  my-attribute: String(Owned("my-value"))
 		 ->  post.attribute: String(Owned("post.value"))
