+++ otel/tests/auto/psr18/psr18-inject.phpt
@@ -60,7 +60,7 @@ Resource
 %A
 Span #0
 	Instrumentation Scope
-		Name         : "php-auto-instrumentation"
+		Name         : "psr18"
 %A
 	Name        : OPTIONS
 	TraceId     : %s
